requirements.txt
Django==1.9
Pillow==3.0.0
Unipath==1.1
django-debug-toolbar==1.4
psycopg2==2.6.1
django_webtest==1.7.8
model_mommy==1.2.6
gunicorn==19.4.1
gevent==1.0.2
django-allauth
django_extensions
ipython
ipdb
-e git+https://github.com/maraujop/django-crispy-forms.git#egg=Package
